Initial support for chronos 3.x for #33
This is the initial draft I wrote for support of chronos >= 3.x
It tries to play nice with versions < 3.x as well and it doesn't keep you from
shooting yourself in the foot. I am undecided at which point you draw the line
in the sand and only allow valid values sent.
Tested vs chronos 2.4 and 3.0.x