Simplify the test environment by a lot
Now it is a matter of passing the appropriate docker node defined in the
docker-compose.yml file. All chronos nodes will expose chronos on port 4400 so
all client calls will be done to localhost port 4400. No need to figure out
magically which port docker-compose is exposing or anything (at least in theory)